feat: [#187871332] update outage bar to be sticky; refactor logo only nav bar 🗺️
