+++ web/src/components/navbar/mobile/NavBarMobileWrapper.tsx
@@ -1,3 +1,4 @@
+import { OutageAlertBar } from "@/components/OutageAlertBar";
 import { NAVBAR_WRAPPER_MOBILE_ID } from "@/lib/utils/helpers";
 import { ReactElement, ReactNode } from "react";
 
@@ -8,14 +9,15 @@ interface NavBarMobileWrapperProps {
 
 export const NavBarMobileWrapper = (props: NavBarMobileWrapperProps): ReactElement => {
   return (
-    <nav
-      aria-label="Primary"
-      className={`width-100 padding-y-05 usa-navbar ${
-        props.scrolled ? "scrolled scrolled-transition bg-white" : ""
-      }`}
-      id={NAVBAR_WRAPPER_MOBILE_ID}
-    >
-      {props.children}
-    </nav>
+    <div className={`${props.scrolled ? "scrolled scrolled-transition bg-white" : ""}`}>
+      <OutageAlertBar />
+      <nav
+        aria-label="Primary"
+        className={`width-100 padding-y-05 usa-navbar `}
+        id={NAVBAR_WRAPPER_MOBILE_ID}
+      >
+        {props.children}
+      </nav>
+    </div>
   );
 };
